LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicyFactory-defaultPolicy-expected.txt


PASS defaultPolicy with no default created is not an error
PASS defaultPolicy returns the correct default policy
PASS defaultPolicy is a read-only property
